Fixed about 150 warnings that showed up in the code when I set the visual studio warning level up to 4 (which I thought it was already on). Now logs the argc and argv. DT is now no longer passed though the updates but is a static variable in the application class. Added background music to the heightmap demo state. Modified the hack in the texturing that got it to work for the heightmap so the data is now deleted in the destructor (which fixes it a bit).
Also added a bunch of my old 2D code so that it can be used with the engine, I will be modified to work with openGL but for the moment it still is built for SDL so will not work with my code. It doesn't have any compile errors however so I feel I can commit it as it is not causing the code to not work while it is there. It is just not used at the moment.